services/goldfin-client/requirements.txt
requests==2.19.1
six==1.11.0
pyyaml==3.13
